pkg/types/config: Break into pkg/validate and pkg/tfvars
This removes a few Cluster properties which are no longer used by the
new installer.  It also drops YAML loading for Cluster now that it's a
one-way map from InstallConfig to Terraform JSON.  I've shifted
validation into its own package, since we should be validating these
at the InstallConfig level.

I've also shifted the Terraform variable generation out of pkg/types,
because it's not about public types.  I've reduced the variable
generation to a single function.

I've dropped the old Cluster (and subtype) specific validation.  We'll
want to add InstallConfig validation once we support loading it from
the disk, but for now we can rely on the UserProvided validation (and
this commit was already large enough ;).
